added: atomic OperationalObjective | added: atomic MarginalFitness

+++ shared_types/src/operational/operational_response_status.rs
@@ -6,11 +6,11 @@ use crate::scheduling_environment::worker_environment::resources::Id;
 pub struct OperationalStatusResponse {
     id: Id,
     number_of_assigned_activities: usize,
-    objective: f64,
+    objective: usize,
 }
 
 impl OperationalStatusResponse {
-    pub fn new(id: Id, number_of_assigned_activities: usize, objective: f64) -> Self {
+    pub fn new(id: Id, number_of_assigned_activities: usize, objective: usize) -> Self {
         Self {
             id,
             number_of_assigned_activities,
